Extracted a common configuration and application codebase for all protocol adapters

Added final to some method parameters
Removed @Configuration annotation from AdapterConfig base class
