+++ adapters/mqtt-vertx/src/main/java/org/eclipse/hono/adapter/mqtt/Config.java
@@ -12,10 +12,8 @@
 
 package org.eclipse.hono.adapter.mqtt;
 
-import io.vertx.core.Vertx;
-import org.eclipse.hono.client.HonoClientConfigProperties;
+import org.eclipse.hono.adapter.AdapterConfig;
 import org.springframework.beans.factory.config.ServiceLocatorFactoryBean;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -23,18 +21,7 @@ import org.springframework.context.annotation.Configuration;
  * Configuration class
  */
 @Configuration
-public class Config {
-
-    private final Vertx vertx = Vertx.vertx();
-
-    @Bean
-    public Vertx getVertx() { return this.vertx; }
-
-    @ConfigurationProperties(prefix = "hono.client")
-    @Bean
-    public HonoClientConfigProperties honoClientConfig() {
-        return new HonoClientConfigProperties();
-    }
+public class Config extends AdapterConfig {
 
     @Bean
     public ServiceLocatorFactoryBean serviceLocator() {
