+++ core/src/main/java/org/eclipse/hono/adapter/AdapterConfig.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Bosch Software Innovations GmbH.
+ * Copyright (c) 2016 Red Hat
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -7,35 +7,31 @@
  * http://www.eclipse.org/legal/epl-v10.html
  *
  * Contributors:
- *    Bosch Software Innovations GmbH - initial creation
- *
+ *    Red Hat - initial creation
  */
 
-package org.eclipse.hono.example;
+package org.eclipse.hono.adapter;
 
+import io.vertx.core.Vertx;
 import org.eclipse.hono.config.HonoClientConfigProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import io.vertx.core.Vertx;
 
 /**
- * Configuration for Example application.
+ * Minimum configuration for protocol adapters
  */
-@Configuration
-public class AppConfiguration {
+public abstract class AdapterConfig {
 
     private final Vertx vertx = Vertx.vertx();
+
+    @Bean
+    public Vertx getVertx() {
+        return vertx;
+    }
 
     @ConfigurationProperties(prefix = "hono.client")
     @Bean
     public HonoClientConfigProperties honoClientConfig() {
         return new HonoClientConfigProperties();
-    }
-
-    @Bean
-    public Vertx vertx() {
-        return vertx;
     }
 }
