+++ device-identities.json
@@ -1,0 +1,7 @@
+[ {
+  "tenant" : "DEFAULT_TENANT",
+  "devices" : [ {
+    "id" : "4711",
+    "data" : { }
+  } ]
+} ]
